moved components and pages code from separate folders to one folder to simplify folder structure
