+++ frontend/src/components/privateRoute.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { Route, Redirect } from 'react-router-dom';
-import AuthContext from '../../context/authContext';
+import AuthContext from '../context/authContext';
 
 const PrivateRoute = ({component: Component, role, ...rest}) => {
   const { role: currentRole } = useContext(AuthContext);
